Restructuring ._connect.pyarrow package
This makes things more natural to use from outside, with or without pyarrow installed.
